+++ src/Tweaks/VoidFieldTweak.cs
@@ -10,6 +10,7 @@ namespace PressureDrop
         {
             if (_hooked) return;
             _hooked = true;
+
             On.RoR2.ArenaMissionController.OnStartServer += ArenaMissionController_OnStartServer;
             On.RoR2.ArenaMissionController.BeginRound += ArenaMissionController_BeginRound;
         }
@@ -18,6 +19,7 @@ namespace PressureDrop
         {
             if (!_hooked) return;
             _hooked = false;
+
             On.RoR2.ArenaMissionController.OnStartServer -= ArenaMissionController_OnStartServer;
             On.RoR2.ArenaMissionController.BeginRound -= ArenaMissionController_BeginRound;
         }